backend/requirements.txt
fastapi
uvicorn[standard]
agno
python-dotenv
requests
google-genai
openai
ddgs
pycoingecko
pandas
numpy
TA-Lib
pandas-ta
ccxt
backtrader
optuna
chromadb
langchain
web3
nltk

